requirements.txt
Pillow>=10.0.0
customtkinter>=5.0.0
tkinterdnd2>=0.3.0
